+++ plugins/plugin.rasp_irc.php
@@ -45,8 +45,8 @@ class PluginRaspIrc extends Plugin {
 
 		$this->setAuthor('undef.de');
 		$this->setVersion('1.0.0');
-		$this->setBuild('2018-05-07');
-		$this->setCopyright('2014 - 2018 by undef.de');
+		$this->setBuild('2019-06-07');
+		$this->setCopyright('2014 - 2019 by undef.de');
 		$this->setDescription('Provides IRC bot to link the server to a channel on an IRC server.');
 
 		$this->registerEvent('onSync',		'onSync');
@@ -285,7 +285,6 @@ class PluginRaspIrc extends Plugin {
 				$this->jukebox_permadd		= $aseco->string2bool($xml['RASP']['JUKEBOX_PERMADD'][0]);
 				$this->jukebox_adminadd		= $aseco->string2bool($xml['RASP']['JUKEBOX_ADMINADD'][0]);
 				$this->jukebox_in_window	= $aseco->string2bool($xml['RASP']['JUKEBOX_IN_WINDOW'][0]);
-				$this->autosave_matchsettings	= $xml['RASP']['AUTOSAVE_MATCHSETTINGS'][0];
 				$this->feature_votes		= $aseco->string2bool($xml['RASP']['FEATURE_VOTES'][0]);
 				$this->prune_records_times	= $aseco->string2bool($xml['RASP']['PRUNE_RECORDS_TIMES'][0]);