+++ plugins/plugin.rasp_votes.php
@@ -58,8 +58,8 @@ class PluginRaspVotes extends Plugin {
 
 		$this->setAuthor('undef.de');
 		$this->setVersion('1.0.0');
-		$this->setBuild('2018-05-07');
-		$this->setCopyright('2014 - 2018 by undef.de');
+		$this->setBuild('2019-06-07');
+		$this->setCopyright('2014 - 2019 by undef.de');
 		$this->setDescription('Provides sophisticated chat-based voting features, similar to (and fully integrated with) MX /add votes.');
 
 		$this->addDependence('PluginRaspJukebox',	Dependence::REQUIRED,	'1.0.0', null);
@@ -1346,7 +1346,6 @@ class PluginRaspVotes extends Plugin {
 				$this->jukebox_permadd		= $aseco->string2bool($xml['RASP']['JUKEBOX_PERMADD'][0]);
 				$this->jukebox_adminadd		= $aseco->string2bool($xml['RASP']['JUKEBOX_ADMINADD'][0]);
 				$this->jukebox_in_window	= $aseco->string2bool($xml['RASP']['JUKEBOX_IN_WINDOW'][0]);
-				$this->autosave_matchsettings	= $xml['RASP']['AUTOSAVE_MATCHSETTINGS'][0];
 				$this->feature_votes		= $aseco->string2bool($xml['RASP']['FEATURE_VOTES'][0]);
 				$this->prune_records_times	= $aseco->string2bool($xml['RASP']['PRUNE_RECORDS_TIMES'][0]);